Refactor UI layouts and extract film detail drawers
